+++ packages/preview-server/src/utils/caniemail/ast/get-used-style-properties.ts
@@ -1,8 +1,8 @@
 /** biome-ignore-all lint/nursery/noNestedComponentDefinitions: There are no components here, just visitor functions */
-import traverse from '@babel/traverse';
+import traverse, { type NodePath } from '@babel/traverse';
+import { inlineStyles, sanitizeStyleSheet } from '@react-email/tailwind';
+import type { StyleSheet } from 'css-tree';
 import type { AST } from '../../../actions/email-validation/check-compatibility';
-import { snakeToCamel } from '../../snake-to-camel';
-import { generateTailwindCssRules } from '../tailwind/generate-tailwind-rules';
 import { getTailwindMetadata } from '../tailwind/get-tailwind-metadata';
 import type { ObjectVariables, SourceLocation } from './get-object-variables';
 
@@ -32,30 +32,36 @@ export const getUsedStyleProperties = async (
   );
 
   if (tailwindMetadata.hasTailwind) {
+    const pathClassNameMap = new Map<string, NodePath>();
+
     traverse(ast, {
       JSXAttribute(path) {
         if (path.node.name.name === 'className') {
           path.traverse({
             StringLiteral(stringPath) {
               const className = stringPath.node.value;
-              const { rules } = generateTailwindCssRules(
-                className.split(' '),
-                tailwindMetadata.context,
-              );
-              for (const rule of rules) {
-                rule.walkDecls((decl) => {
-                  styleProperties.push({
-                    location: stringPath.node.loc,
-                    name: snakeToCamel(decl.prop),
-                    value: decl.value,
-                  });
-                });
-              }
+              pathClassNameMap.set(className, stringPath);
+              const candidates = className.split(/\s+/);
+              tailwindMetadata.tailwindSetup.addUtilities(candidates);
             },
           });
         }
       },
     });
+    const styleSheet =
+      tailwindMetadata.tailwindSetup.getStyleSheet() as StyleSheet;
+    sanitizeStyleSheet(styleSheet);
+
+    for (const [className, nodePath] of pathClassNameMap.entries()) {
+      const styles = inlineStyles(styleSheet, className.split(/\s+/));
+      for (const [name, value] of Object.entries(styles)) {
+        styleProperties.push({
+          location: nodePath.node.loc,
+          name,
+          value,
+        });
+      }
+    }
   }
 
   traverse(ast, {